+++ packages/headers-parser/tests/helpers/main.ts
@@ -1,24 +1,32 @@
 import { fileURLToPath } from 'url'
 
-import { parseAllHeaders } from '../../src/index.js'
+import { type MinimalHeader, parseAllHeaders } from '../../src/index.js'
 
 const FIXTURES_DIR = fileURLToPath(new URL('../fixtures', import.meta.url))
 
-export const parseHeaders = async function ({ headersFiles, netlifyConfigPath, configHeaders, ...input }: any) {
+export const parseHeaders = async function ({
+  headersFiles,
+  netlifyConfigPath,
+  configHeaders,
+  minimal,
+}: {
+  headersFiles: undefined | string[]
+  netlifyConfigPath: undefined | string
+  configHeaders: undefined | MinimalHeader[]
+  minimal: boolean
+}) {
   return await parseAllHeaders({
     ...(headersFiles && { headersFiles: headersFiles.map(addFileFixtureDir) }),
     ...(netlifyConfigPath && { netlifyConfigPath: addConfigFixtureDir(netlifyConfigPath) }),
     configHeaders,
-    // Default `minimal` to `true` but still allows passing `undefined` to
-    // test the default value of that option
-    minimal: 'minimal' in input ? input.minimal : true,
+    minimal,
   })
 }
 
-const addFileFixtureDir = function (name) {
+const addFileFixtureDir = function (name: string): string {
   return `${FIXTURES_DIR}/headers_file/${name}`
 }
 
-const addConfigFixtureDir = function (name) {
+const addConfigFixtureDir = function (name: string): string {
   return `${FIXTURES_DIR}/netlify_config/${name}.toml`
 }