+++ packages/zip-it-and-ship-it/tests/v2api.test.ts
@@ -342,7 +342,7 @@ describe.runIf(semver.gte(nodeVersion, '18.13.0'))('V2 functions API', () => {
       },
     })
 
-    expect(files[0].runtimeVersion).toBe('nodejs18.x')
+    expect(files[0].runtimeVersion).toBe('nodejs22.x')
   })
 
   test('Returns Node.js 18 if invalid version is set', async () => {
@@ -357,7 +357,7 @@ describe.runIf(semver.gte(nodeVersion, '18.13.0'))('V2 functions API', () => {
       },
     })
 
-    expect(files[0].runtimeVersion).toBe('nodejs18.x')
+    expect(files[0].runtimeVersion).toBe('nodejs22.x')
   })
 
   test('Returns no Node.js version if version is newer than 18 but not a valid runtime', async () => {
@@ -714,11 +714,6 @@ describe.runIf(semver.gte(nodeVersion, '18.13.0'))('V2 functions API', () => {
       const fixtureName = 'v2-api'
       const { files } = await zipFixture(fixtureName, {
         fixtureDir: FIXTURES_ESM_DIR,
-        opts: {
-          featureFlags: {
-            zisi_add_metadata_file: true,
-          },
-        },
       })
       const [unzippedFunction] = await unzipFiles(files)
       const bootstrapPath = getBootstrapPath()
@@ -735,9 +730,6 @@ describe.runIf(semver.gte(nodeVersion, '18.13.0'))('V2 functions API', () => {
         fixtureDir: FIXTURES_ESM_DIR,
         opts: {
           branch: 'main',
-          featureFlags: {
-            zisi_add_metadata_file: true,
-          },
         },
       })
       const [unzippedFunction] = await unzipFiles(files)
@@ -765,9 +757,6 @@ describe.runIf(semver.gte(nodeVersion, '18.13.0'))('V2 functions API', () => {
     const { files } = await zipFixture('v2-api', {
       fixtureDir: FIXTURES_ESM_DIR,
       opts: {
-        featureFlags: {
-          zisi_add_metadata_file: true,
-        },
         manifest: manifestPath,
       },
     })